Construct an S3 object of type stansim_simulation

A constructor function for creating S3 objects of class stansim_simulation. stansim_simulation objects are the basic unit of output from calls to the fit_models() function and collects the specified data for all stan models fitted.

stansim_simulation(sim_name, stansim_uni_list, start_time, end_time, raw_call,
  seed)

Arguments

sim_name
The name to be given to the simulation represented by the stansim_simulation object.

stansim_uni_list
A list of objects with S3 class stan_sim_uni. This is an unexported class used to store the outcomes of individual simulation runs internal to the fit_models() function.

start_time
System time when fit_models() was called.

end_time
System time when the results from fit_models() were returned.

raw_call
The values of all arguments provided to fit_models() when first ran. This is used for any refitting of datasets using the refit() method.

seed
The global seed for the fit_models() call.

Value

An S3 object of class stansim_simulation recording relevant simulation data.